wip! platform/surface: aggregator: Allow devices to be marked as hot-removed
Some SSAM devices, notably keyboard and touchpad on the Surface Book 3
and Surface Pro X, can be hot-removed. When this occurs, communication
with the device may fail and time out. This timeout may unnecessarily
block and slow down device removal and even cause issues when the
devices are detached and re-attached quickly. Thus, communication should
generally be avoided once hot-removal is detected.

While we already remove a device as soon as we detect its hot-removal,
the corresponding device driver may still attempt to communicate with
the device during teardown. This is especially critical as communication
failure may also extend to event disablement.

Add a flag to allow marking devices as hot-removed. This can then be
used during client driver teardown to check if any communication
attemtps should be avoided.