scsi: sd: Enable modern protocol features on more devices
Due to legacy USB devices having a tendency to either lock up or return
garbage if one attempts to query device capabilities, the USB transport
disables VPD pages by default. This prevents discard from being properly
configured on most modern USB-attached SSDs.

Introduce two additional heuristics to determine whether VPD pages should
be consulted.

The first heuristic fetches VPD pages if a device reports that Logical
Block Provisioning is enabled. It is very unusual for a device to support
thin provisioning and not provide the associated VPDs.  Consequently, if a
device reports that Logical Block Provisioning is enabled (LBPME) in READ
CAPACITY(16) response, the scsi_device has no VPDs attached, and the
reported SPC version is larger than 3, then an attempt will be made to read
the VPD pages during revalidate.

The second heuristic relies on the fact that almost all modern devices
return a set of version descriptors in the INQUIRY response. These
descriptors outline which version of various protocol features are
supported. If a device manufacturer has gone through the effort of filling
out compliance descriptors, it is highly unlikely that VPD pages are not
supported. So if a device provides version descriptors in the INQUIRY
response, the scsi_device has no VPDs attached, and the reported SBC
version is larger than 2, then an attempt will be made to read the VPD
pages. In addition, READ CAPACITY(16) will be preferred over READ
CAPACITY(10) to facilitate accessing the LBPME flag.

The benefit to relying on INQUIRY is that it is data we already have. We do
not have to blindly poke the device for additional information and risk
confusing it.

Extracting the SBC version is done by a new helper, sd_sbc_version().
Another helper is provided to determine whether a scsi_device has VPD pages
attached or not.

Link: https://lore.kernel.org/r/20220302053559.32147-15-martin.petersen@oracle.com
Reported-by: Aman Karmani <aman@tmm1.net>
Tested-by: Aman Karmani <aman@tmm1.net>
Reported-by: David Sebek <dasebek@gmail.com>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Bart Van Assche <bvanassche@acm.org>
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>