wip! HID: spi-hid: Fix NULL-pointer exception when device is not initialized
The device will not be initialized when it is not properly reset. In
that case, normal input data may be received, which causes the driver to
crash as buffers are not allocated yet.

Fix this by checking if the device is ready and ignore any input data if
it is not.