media: i2c: Check fwnode->secondary for endpoint
The ov8865 driver is one of those that can be connected to a CIO2
device by the cio2-bridge code. This means that the absence of an
endpoint for this device is not necessarily fatal, as one might be
built by the cio2-bridge when it probes. Check fwnode->secondary for
an endpoint, and defer probing if one isn't found rather than fail.

Signed-off-by: Daniel Scally <djrscally@gmail.com>