platform/surface: aggregator: reconfigure GPIO interrupt on probe
On system initialization the IRQ may be preconfigured by gpiolib-acpi as
per the settings defined in the machine's ACPI table.  To deconfigure
this interrupt as being set as a wake source for Si0x and S3 sleep
states we need to call disable_irq_wake, however there is some error
checking here to ensure the calls are balanced, so we enable the wakeup
source first, before following up with disabling.

Signed-off-by: Sachi King <nakato@nakato.io>