platform/surface: aggregator_registry: Remove base hub driver
The base hub was a virtual device hub for Surface System Aggregator
Module (SSAM) client devices contained in the detachable Surface Book 3
base. Remove it as it is no longer needed.

In the previous change, we have moved all devices from the base hub to
the KIP hub. That change has also removed the only base-hub-device that
ever existed, as it was essentially replaced by the KIP-hub-device and
thus was no longer needed. This means that there is no remaining
hub-device against which the base hub driver can load, i.e. it is now
essentially dead code. So remove the driver as well.

Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>
Patchset: surface-sam