scsi: sd: Move WRITE_ZEROES configuration to a separate function
In preparation for adding support for the WRITE SAME(16) NDOB flag, move
configuration of the WRITE_ZEROES operation to a separate function. This is
done to facilitate fetching all VPD pages before choosing the appropriate
zeroing method for a given device.

The deferred configuration also allows us to mirror the discard behavior
and permit the user to revert a device to the kernel default configuration
by echoing "default" to the sysfs file.

Link: https://lore.kernel.org/r/20220302053559.32147-11-martin.petersen@oracle.com
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>