rust: avoid all GCC plugins, not just the randstruct one
We can keep support for some of the GCC plugins. However, it is simpler
and safer to just disable GCC plugin support since they may be going
away anyway [0]. Furthermore, if they are kept, new or future ones
could also break Rust support in subtle ways. And, in any case, Rust
support with GCC-built kernels is very experimental. So, at least for
the moment, let's focus on other things.

[0] https://lwn.net/Articles/851090/

Signed-off-by: Miguel Ojeda <ojeda@kernel.org>