PCI: qcom: Add system PM support
Add suspend and resume callbacks to handle system suspend and resume in
the Qcom PCIe controller driver. When the system suspends, PME turnoff
message will be sent to the device and the RC driver will wait for the
device to enter L23 Ready state. After that, the PHY will be powered down
and clocks/regulators will be disabled.

When the system resumes, RC driver gets initialized and re-establishes the
link.

The system suspend and resume are only handled when the underlying platform
supports the "suspend_poweroff" flag that notifies the PCI drivers about
powerdown during suspend.

Otherwise, these callbacks are NO-OPs. Currently, this flag is only enabled
for SC7280 SoC, where aggressive powersaving is required during system
suspend. It should be noted that on this platform, RPMh will also cutoff
the power to PCIe domain during suspend, independent of the RC driver as
rest of the RPMh clients de-vote for the resources. But ideally, the PCIe
devices should be put into D3cold state during suspend by the PCIe RC
driver and that's what this commit also does.

Signed-off-by: Prasad Malisetty <quic_pmaliset@quicinc.com>
[mani: Reworded commit message, removed pipe_clk & added "suspend_poweroff" flag]
Signed-off-by: Manivannan Sadhasivam <manivannan.sadhasivam@linaro.org>