scsi: iscsi: Add helper to remove a session from the kernel
qedi requires that we at least tell the FW to disconnect and clean up
connections during shutdown. Commit d1f2ce77638d ("scsi: qedi: Fix host
removal with running sessions") converted the driver to use the libicsi
helper to drive session removal.  The problem is that during shutdown
iscsid will not be running so when we try to remove the root session we
will hang wait for userspace to reply.

This patch adds a helper that will drive the destruction of sessions like
these during system shutdown.

Link: https://lore.kernel.org/r/20220519003518.34187-3-michael.christie@oracle.com
Reviewed-by: Lee Duncan <lduncan@suse.com>
Signed-off-by: Mike Christie <michael.christie@oracle.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>