wip! arm64: dts: surface-pro-x: Use 'enable-gpios' instead of gpio-based voltage regulator for panel
It turns out that the voltage regulator is power-cycled during boot,
which in turn cycles the panel-vdd-enable GPIO pin. The SAM EC notices
that the panel-vdd-enable GPIO pin is driven high by the SOC and takes
this as signal to relinquish control (drive signals by EC and SOC are
or-ed together). Since we are, however, not actually trying to drive the
panel but just power-cycling the regulator, we turn off the panel,
leading to a loss of display during boot.

If we use the panel-vdd-enable GPIO as the panel enable gpio instead of
using it inside a voltage regulator, we can avoid this cycling behavior.
Therefore, we let SAM drive the panel for a bit longer until we are
actually ready to take over.